# namviet-backend-core/docs/swagger.yaml
definitions:
...
  ai_agent_memories.AIAgentMemory:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key:
        type: string
      session_id:
        type: string
      user_id:
        type: string
      value:
        items:
          type: string
        type: array
    type: object
  ai_agent_memories.CreateAIAgentMemoryRequest:
    properties:
      expires_at:
        type: string
      key:
        type: string
      session_id:
        type: string
      user_id:
        type: string
      value:
        items:
          type: string
        type: array
    required:
    - key
    - value
    type: object
  ai_agent_memories.UpdateAIAgentMemoryRequest:
    properties:
      expires_at:
        type: string
      key:
        type: string
      session_id:
        type: string
      user_id:
        type: string
      value:
        items:
          type: string
        type: array
    type: object
...
  categories.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  categories.CreateCategoryRequest:
    properties:
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      status:
        type: string
    required:
    - name
    - slug
    type: object
  categories.UpdateCategoryRequest:
    properties:
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for Nam Viet ERP System
  title: Nam Viet ERP API
  version: "1.0"
paths:
...
  /ai-agent-memories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all ai agent memories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ai_agent_memories.AIAgentMemory'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All AI Agent Memories
      tags:
      - AI Ecosystem
    post:
      consumes:
      - application/json
      description: Create a new ai agent memory
      parameters:
      - description: Memory Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai_agent_memories.CreateAIAgentMemoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ai_agent_memories.AIAgentMemory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new AI Agent Memory
      tags:
      - AI Ecosystem
  /ai-agent-memories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an ai agent memory
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete AI Agent Memory
      tags:
      - AI Ecosystem
    get:
      consumes:
      - application/json
      description: Retrieve a specific ai agent memory
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai_agent_memories.AIAgentMemory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Agent Memory by ID
      tags:
      - AI Ecosystem
    put:
      consumes:
      - application/json
      description: Update ai agent memory details
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai_agent_memories.UpdateAIAgentMemoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai_agent_memories.AIAgentMemory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update AI Agent Memory
      tags:
      - AI Ecosystem
...
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/categories.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Category Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/categories.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/categories.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Category
      tags:
      - Categories
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categories.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update category details
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/categories.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categories.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Category
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"